fix: rename rate_limiter package to ratelimiter to satisfy revive linter
- Rename pkg/rate_limiter/ directory to pkg/ratelimiter/
- Change package declaration from rate_limiter to ratelimiter
- Update all import paths and usages across modes.go, main.go, modes_test.go, integration_test.go
- Reformat DoScanTableWithConfig signature to fix golines line-length violation
