+++ modes.go
@@ -19,7 +19,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/scylladb/scylla-bench/internal/clock"
-	"github.com/scylladb/scylla-bench/pkg/rate_limiter"
+	"github.com/scylladb/scylla-bench/pkg/ratelimiter"
 	"github.com/scylladb/scylla-bench/pkg/results"
 	"github.com/scylladb/scylla-bench/pkg/workloads"
 	"github.com/scylladb/scylla-bench/random"
@@ -148,7 +148,7 @@ var (
 func RunConcurrently(
 	clk clock.Clock,
 	maximumRate int,
-	workload func(id int, testResult *results.TestThreadResult, rateLimiter rate_limiter.RateLimiter),
+	workload func(id int, testResult *results.TestThreadResult, rateLimiter ratelimiter.RateLimiter),
 ) *results.TestResults {
 	var timeOffsetUnit int64
 	if maximumRate != 0 {
@@ -166,7 +166,7 @@ func RunConcurrently(
 		testResult := totalResults.GetTestResult(i)
 		go func(i int) {
 			timeOffset := time.Duration(timeOffsetUnit * int64(i))
-			workload(i, testResult, rate_limiter.NewRateLimiter(clk, maximumRate, timeOffset))
+			workload(i, testResult, ratelimiter.NewRateLimiter(clk, maximumRate, timeOffset))
 		}(i)
 	}
 
@@ -227,7 +227,7 @@ func RunTest(
 	config ExecutionConfig,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	test func(rb *results.TestThreadResult) (time.Duration, error),
 ) {
 	config = config.normalized()
@@ -526,7 +526,7 @@ func DoWrites(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	RunTest(
@@ -543,7 +543,7 @@ func DoWritesWithConfig(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	RunTest(
@@ -560,7 +560,7 @@ func DoBatchedWritesWithConfig(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	config = config.normalized()
@@ -650,7 +650,7 @@ func DoBatchedWrites(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	DoBatchedWritesWithConfig(DefaultExecutionConfig(), session, threadResult, workload, rateLimiter, validateData)
@@ -660,7 +660,7 @@ func DoCounterUpdates(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	_ bool,
 ) {
 	DoCounterUpdatesWithConfig(DefaultExecutionConfig(), session, threadResult, workload, rateLimiter, false)
@@ -671,7 +671,7 @@ func DoCounterUpdatesWithConfig(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	_ bool,
 ) {
 	config = config.normalized()
@@ -720,7 +720,7 @@ func DoReads(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	config := DefaultExecutionConfig()
@@ -731,7 +731,7 @@ func DoCounterReads(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	config := DefaultExecutionConfig()
@@ -926,7 +926,7 @@ func DoReadsFromTableWithConfig(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	RunTest(config, threadResult, workload, rateLimiter, createReadTestFuncWithConfig(config, table, session, workload, validateData))
@@ -937,13 +937,20 @@ func DoReadsFromTable(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	DoReadsFromTableWithConfig(DefaultExecutionConfig(), table, session, threadResult, workload, rateLimiter, validateData)
 }
 
-func DoScanTableWithConfig(config ExecutionConfig, session *gocql.Session, threadResult *results.TestThreadResult, workload workloads.Generator, rateLimiter rate_limiter.RateLimiter, _ bool) {
+func DoScanTableWithConfig(
+	config ExecutionConfig,
+	session *gocql.Session,
+	threadResult *results.TestThreadResult,
+	workload workloads.Generator,
+	rateLimiter ratelimiter.RateLimiter,
+	_ bool,
+) {
 	config = config.normalized()
 	request := fmt.Sprintf("SELECT * FROM %s.%s WHERE token(pk) >= ? AND token(pk) <= ?", config.KeyspaceName, config.TableName)
 
@@ -978,7 +985,7 @@ func DoScanTableWithConfig(config ExecutionConfig, session *gocql.Session, threa
 	})
 }
 
-func DoScanTable(session *gocql.Session, threadResult *results.TestThreadResult, workload workloads.Generator, rateLimiter rate_limiter.RateLimiter, validateData bool) {
+func DoScanTable(session *gocql.Session, threadResult *results.TestThreadResult, workload workloads.Generator, rateLimiter ratelimiter.RateLimiter, validateData bool) {
 	DoScanTableWithConfig(DefaultExecutionConfig(), session, threadResult, workload, rateLimiter, validateData)
 }
 
@@ -987,7 +994,7 @@ func DoMixedWithConfig(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	config = config.normalized()
@@ -1031,7 +1038,7 @@ func DoMixed(
 	session *gocql.Session,
 	threadResult *results.TestThreadResult,
 	workload workloads.Generator,
-	rateLimiter rate_limiter.RateLimiter,
+	rateLimiter ratelimiter.RateLimiter,
 	validateData bool,
 ) {
 	DoMixedWithConfig(DefaultExecutionConfig(), session, threadResult, workload, rateLimiter, validateData)
